Add List of Items

+++ MyShop/Portion.cs
@@ -5,7 +5,7 @@ namespace MyShop
     [Serializable]
     public class Portion
     {
-        public double Amount { set; get; }
+        public decimal Amount { set; get; }
         public Item Item { set; get; }
     }
 }
